# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "销售赋能中心后端 API 文档",
        "title": "TaishanXD V2 API",
        "contact": {},
        "version": "2.0.0"
    },
    "host": "localhost:4007",
    "basePath": "/",
    "paths": {
...
        "/api/wecom/kefu-messages": {
            "get": {
                "description": "根据 external_userid 查询微信客服消息，包含用户提问（received）和机器人回复（sent），支持时间范围筛选",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "企业微信"
                ],
                "summary": "查询客服聊天记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "外部用户ID",
                        "name": "external_userid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间（RFC3339 格式，如 2026-05-01T00:00:00Z）",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "截止时间（RFC3339 格式）",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "返回条数（最大 500）",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.KefuMessage"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/wecom/messages": {
            "get": {
                "description": "根据企业微信 platform_user_id 查询该用户相关的所有消息（作为发送者或接收者），支持时间范围和会话类型筛选",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "企业微信"
                ],
                "summary": "查询用户聊天记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "企业微信用户ID（如 GuoTongJia）",
                        "name": "platform_user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MongoDB 公司 ID（可选，不传则不过滤公司）",
                        "name": "company_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "起始时间（毫秒时间戳）",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "截止时间（毫秒时间戳）",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "all",
                        "description": "会话类型：all / group / private",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "返回条数（最大 500）",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.WecomMessage"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.KefuMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "direction": {
                    "type": "string"
                },
                "externalUserId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "messageId": {
                    "type": "string"
                },
                "msgType": {
                    "type": "string"
                },
                "openKfId": {
                    "type": "string"
                }
            }
        },
...
        "model.WecomMessage": {
            "type": "object",
            "properties": {
                "chatId": {
                    "type": "string"
                },
                "chatName": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createTime": {
                    "type": "integer"
                },
                "direction": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "messageId": {
                    "type": "string"
                },
                "msgType": {
                    "type": "string"
                },
                "receiveId": {
                    "type": "string"
                },
                "receiveIdType": {
                    "type": "string"
                },
                "senderId": {
                    "type": "string"
                },
                "senderIdType": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}